requirements.txt
Flask
fpdf
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread
matplotlib
oauth2client
pdf2image
pdfrw
Pillow
pymongo
